+++ R/globals.R
@@ -4,7 +4,5 @@ globalVariables(unique(c(
   # add_fun_to_parsed
   "fun_name", "rox_filename", "chunk_filename", ".",
   # get_functions
-  ".",
-  # split_rmd
-  "type", "lines", "title"
+  "."
 )))